+++ apps/ledger-live-mobile/src/analytics/segment.ts
@@ -34,6 +34,7 @@ import {
   notificationsSelector,
   knownDeviceModelIdsSelector,
   customImageTypeSelector,
+  userNpsSelector,
 } from "../reducers/settings";
 import { knownDevicesSelector } from "../reducers/ble";
 import { DeviceLike, State } from "../reducers/types";
@@ -125,6 +126,7 @@ const extraProperties = async (store: AppStore) => {
     : [];
   const hasGenesisPass = hasNftInAccounts(GENESIS_PASS_COLLECTION_CONTRACT, accounts);
   const hasInfinityPass = hasNftInAccounts(INFINITY_PASS_COLLECTION_CONTRACT, accounts);
+  const nps = userNpsSelector(state);
 
   return {
     appVersion,
@@ -160,6 +162,7 @@ const extraProperties = async (store: AppStore) => {
     staxDeviceUser: knownDeviceModelIds.stax,
     staxLockscreen: customImageType || "none",
     ...getFeatureFlagProperties(),
+    nps,
   };
 };
 
@@ -364,7 +367,7 @@ export const screen = async (
    */
   avoidDuplicates?: boolean,
 ) => {
-  const fullScreenName = category + (name ? ` ${name}` : "");
+  const fullScreenName = (category || "") + (category && name ? " " : "") + (name || "");
   const eventName = `Page ${fullScreenName}`;
   if (avoidDuplicates && eventName === lastScreenEventName.current) return;
   lastScreenEventName.current = eventName;