+++ apps/ledger-live-mobile/src/screens/PostOnboarding/PostOnboardingHub.tsx
@@ -1,18 +1,8 @@
 import React, { useCallback, useEffect, useRef } from "react";
-import { Divider, Flex, IconsLegacy, Log, Text } from "@ledgerhq/native-ui";
+import { Divider, Flex, Text, Button } from "@ledgerhq/native-ui";
 import { useTranslation } from "react-i18next";
-import { ScrollView, StyleSheet } from "react-native";
-import { useFocusEffect } from "@react-navigation/native";
+import { ScrollView } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import Animated, {
-  cancelAnimation,
-  interpolate,
-  runOnJS,
-  useAnimatedStyle,
-  useSharedValue,
-  withDelay,
-  withTiming,
-} from "react-native-reanimated";
 import {
   useAllPostOnboardingActionsCompleted,
   usePostOnboardingHubState,
@@ -29,8 +19,6 @@ import { PostOnboardingNavigatorParamList } from "../../components/RootNavigator
 import { useCompleteActionCallback } from "../../logic/postOnboarding/useCompleteAction";
 import { track, TrackScreen } from "../../analytics";
 import Link from "../../components/wrappedUi/Link";
-
-const AnimatedFlex = Animated.createAnimatedComponent(Flex);
 
 type NavigationProps = BaseComposite<
   StackNavigatorProps<PostOnboardingNavigatorParamList, ScreenName.PostOnboardingHub>
@@ -58,12 +46,17 @@ const PostOnboardingHub = ({ navigation, route }: NavigationProps) => {
 
   useEffect(
     /**
-     * Complete claim NFT action if the route param completed is true
+     * Complete claim NFT action if the route param completed is true (used for NFT redeem)
+     * Or complete all steps (For debud purposes)
      * */
     () => {
       if (route?.params?.completed === "true") {
         track("deeplink", { action: "Claim NFT return to setup" });
         completePostOnboardingAction(PostOnboardingActionId.claimNft);
+      }
+      if (route?.params?.allCompleted === "true") {
+        completePostOnboardingAction(PostOnboardingActionId.claimNft);
+        completePostOnboardingAction(PostOnboardingActionId.customImage);
       }
     },
     [clearLastActionCompleted, completePostOnboardingAction, route],
@@ -78,76 +71,7 @@ const PostOnboardingHub = ({ navigation, route }: NavigationProps) => {
     });
   }, [navigation]);
 
-  /**
-   * At 0: regular screen
-   * At 0.5: opaque black/white backdrop drawn on top
-   * At 1: "finished" text drawn on top of backdrop
-   */
-  const animDoneValue = useSharedValue(0);
-
   const allDone = useAllPostOnboardingActionsCompleted();
-
-  const animationTimeout = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const clearAnimationTimeout = useCallback(() => {
-    !allDone && animationTimeout.current && clearTimeout(animationTimeout.current);
-  }, [allDone]);
-
-  const triggerEndAnimation = useCallback(() => {
-    const onAnimEnd = () => {
-      clearAnimationTimeout();
-      animationTimeout.current = setTimeout(() => {
-        navigateToMainScreen();
-      }, 3000);
-    };
-    animDoneValue.value = withDelay(
-      2000,
-      withTiming(1, { duration: 1500 }, finished => {
-        if (finished) {
-          runOnJS(onAnimEnd)();
-        }
-      }),
-    );
-
-    /**
-     * Preventing screen closing:
-     * - removing header right (close button)
-     * - disabling gestures (necessary for iOS)
-     * - adding listener to "beforeRemove" (necessary for Android as the back button would still close otherwise)
-     */
-    navigation.setOptions({ gestureEnabled: false, headerRight: () => null });
-    navigation.getParent()?.setOptions({ gestureEnabled: false });
-    allowClosingScreen.current = false;
-    const listenerCleanup = navigation.addListener("beforeRemove", e => {
-      if (!allowClosingScreen.current) e.preventDefault();
-    });
-    return () => {
-      listenerCleanup();
-      clearAnimationTimeout();
-      cancelAnimation(animDoneValue);
-    };
-  }, [clearAnimationTimeout, navigation, navigateToMainScreen, animDoneValue, animationTimeout]);
-
-  useEffect(() => clearAnimationTimeout);
-
-  const triggerEndAnimationWrapped = useCallback(() => {
-    if (allDone) triggerEndAnimation();
-  }, [allDone, triggerEndAnimation]);
-
-  useFocusEffect(triggerEndAnimationWrapped);
-
-  const doneContainerStyle = useAnimatedStyle(
-    () => ({
-      opacity: interpolate(animDoneValue.value, [0, 0.5, 1], [0, 1, 1]),
-    }),
-    [animDoneValue],
-  );
-
-  const doneContentStyle = useAnimatedStyle(
-    () => ({
-      opacity: interpolate(animDoneValue.value, [0, 0.5, 1], [0, 0, 1]),
-    }),
-    [animDoneValue],
-  );
 
   const productName = getDeviceModel(deviceModelId || DeviceModelId.nanoX)?.productName;
 
@@ -168,23 +92,37 @@ const PostOnboardingHub = ({ navigation, route }: NavigationProps) => {
         flex={1}
         paddingBottom={safeAreaInsets.bottom}
       >
-        <Text variant="h1Inter" fontWeight="semiBold" mb={8}>
-          {allDone
-            ? t("postOnboarding.hub.allDoneTitle", {
-                productName,
-              })
-            : t("postOnboarding.hub.title", { productName })}
-        </Text>
-        <ScrollView>
-          {actionsState.map((action, index, arr) => (
-            <React.Fragment key={index}>
-              <PostOnboardingActionRow {...action} />
-              {index !== arr.length - 1 && <Divider />}
-            </React.Fragment>
-          ))}
-        </ScrollView>
-        <Flex mt={8} alignItems="center" justifyContent="center">
-          {allDone ? null : (
+        <Flex>
+          <Flex pb={8}>
+            <Text variant="h1Inter" fontWeight="semiBold">
+              {allDone
+                ? t("postOnboarding.hub.allDoneTitle", {
+                    productName,
+                  })
+                : t("postOnboarding.hub.title", { productName })}
+            </Text>
+          </Flex>
+          <ScrollView>
+            {actionsState.map((action, index, arr) => (
+              <React.Fragment key={index}>
+                <PostOnboardingActionRow {...action} />
+                {index !== arr.length - 1 && <Divider />}
+              </React.Fragment>
+            ))}
+          </ScrollView>
+        </Flex>
+        <Flex my={7} alignItems="center" justifyContent="center">
+          {allDone ? (
+            <Button
+              alignSelf="stretch"
+              size="large"
+              type="main"
+              outline={false}
+              onPress={navigateToMainScreen}
+            >
+              {t("postOnboarding.hub.viewWallet")}
+            </Button>
+          ) : (
             <Link
               size="large"
               onPress={navigateToMainScreen}
@@ -196,23 +134,6 @@ const PostOnboardingHub = ({ navigation, route }: NavigationProps) => {
           )}
         </Flex>
       </Flex>
-      {allDone && (
-        <AnimatedFlex
-          style={[doneContainerStyle, StyleSheet.absoluteFillObject]}
-          justifyContent="center"
-          alignItems="center"
-        >
-          <Flex backgroundColor="background.main" style={StyleSheet.absoluteFillObject} />
-          <AnimatedFlex style={doneContentStyle}>
-            <Flex flexDirection="column" alignItems="center" p={8}>
-              <IconsLegacy.CircledCheckSolidMedium color="success.c50" size={54} />
-              <Flex height={83} />
-              <Log>{t("postOnboarding.hub.done")}</Log>
-              <Flex height={100} />
-            </Flex>
-          </AnimatedFlex>
-        </AnimatedFlex>
-      )}
     </>
   );
 };